Consistent trainval splits
Before, trainval splits were calculated randomly, which was detrimental if using these scripts multiple times on the same dataset, as it would mix up the train/val images and pollute the validity of mAP, as the model already trained on the mixed up images. I also added more checks to minimize confusion when utilizing scripts.